refactor: remove tabs demo module
