+++ vaadin-tabs-flow-parent/pom.xml
@@ -12,7 +12,6 @@
     <modules>
         <module>vaadin-tabs-flow</module>
         <module>vaadin-tabs-testbench</module>
-        <module>vaadin-tabs-flow-demo</module>
     </modules>
     <profiles>
         <profile>
